// ui/adminv2/apps/web/types/ProxyServerStatus.ts
/*
 * Copyright (c) 2026. Mikhail Kulik.
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface ProxyServerStatus {
  readonly running: boolean
  config: ProxyServerConfig
}
export interface ProxyServerConfig {
  listen_port: number
  read_header_timeout: number
  tls_enabled: boolean
  readonly tls_cert_path?: string
  readonly tls_key_path?: string
  readonly tls_cert_hash?: string
  readonly tls_key_hash?: string
}